remove rir analysis/optimization stuff
