+++ rir/src/ir/Optimizer.cpp
@@ -1,7 +1,5 @@
 #include "ir/Optimizer.h"
-#include "optimization/cleanup.h"
-#include "optimization/localize.h"
-#include "optimization/stupid_inline.h"
+#include "ir/cleanup.h"
 
 namespace rir {
 
@@ -20,30 +18,14 @@ bool Optimizer::optimize(CodeEditor& code, int steam) {
     return changed;
 }
 
-bool Optimizer::inliner(CodeEditor& code, bool stable) {
-    Localizer local(code, stable);
-    local.run();
-    bool changed = code.changed;
-    if (code.changed)
-        code.commit();
-    StupidInliner inl(code);
-    inl.run();
-    changed = changed || code.changed;
-    if (code.changed)
-        code.commit();
-    return changed;
-}
-
 SEXP Optimizer::reoptimizeFunction(SEXP s) {
     Function* fun = Function::unpack(s);
-    bool safe = !fun->envLeaked && !fun->envChanged;
 
     CodeEditor code(s);
 
     for (int i = 0; i < 16; ++i) {
-        bool changedInl = Optimizer::inliner(code, safe);
         bool changedOpt = Optimizer::optimize(code, 8);
-        if (!changedInl && !changedOpt) {
+        if (!changedOpt) {
             if (i == 0)
                 return nullptr;
             break;
