+++ rir/src/ir/Optimizer.h
@@ -8,7 +8,6 @@ namespace rir {
 class Optimizer {
   public:
     static bool optimize(CodeEditor&, int steam = 10);
-    static bool inliner(CodeEditor&, bool stableEnv);
     static SEXP reoptimizeFunction(SEXP);
 };
 }
