--- rir/src/optimization/cp.cpp
@@ -1,7 +1,0 @@
-#include "cp.h"
-
-namespace rir {
-
-SEXP const CP_Value::bottom_ = nullptr;
-SEXP const CP_Value::top_ = (SEXP)(1);
-}
